fix(migrations): the schema advisory lock excluded nothing — every instance computed a different key
The migration lock key was Math.Abs(schema.GetHashCode()) % int.MaxValue. .NET seeds
string hashing randomly per process, so each instance computed its own key, each
acquired its own private lock, and all of them entered the migration path together.
The lock had no effect on the thing it exists to serialize.

What still saved it: the in-lock hash re-check plus IF NOT EXISTS. Neither covers
CREATE OR REPLACE FUNCTION or settings-gated data migrations, so concurrent DDL was
reachable, not merely theoretical. The same expression also threw OverflowException
whenever GetHashCode returned int.MinValue — a startup crash decided by which hash
seed a process happened to draw.

The framework had already solved this once: CollectiveApplyLockKey uses FNV-1a 64
precisely because GetHashCode is per-process randomized. The migration lock never got
the same treatment. Extract that hash to a shared Fnv1a64 and give schema
initialization its own namespaced key, so the two families cannot collide in Postgres's
single-bigint key space.

- SchemaInitializationLockKey.Compute(schema), keyed "wh_schema_init:<schema>".
  Normalizes null/empty to "public", matching _transformMigrationSql — otherwise two
  instances of one service, one configured with an explicit "public" and one with none,
  take different locks while migrating the same physical schema.
- pg_try_advisory_xact_lock takes bigint, so there is nothing to truncate and no
  Math.Abs to overflow. SchemaInitializationLog's lockId params widen int -> long.
- CollectiveApplyLockKey now delegates to the shared hash; pinned-value tests prove the
  refactor changed none of its keys.

Tests pin the key to literals derived from the FNV-1a definition rather than captured
from a run: a per-process value cannot match a fixed literal, and a value that drifts
between library versions would break a rolling deploy just as badly. The integration
test holds the key computed from the published function — standing in for the other
process, since two initializers inside one test process would agree even on the broken
code — forces the slow path, and asserts the initializer is excluded rather than
sailing through.

+++ src/Whizbang.Data.EFCore.Postgres/SchemaInitializationLog.cs
@@ -17,22 +17,22 @@ public static partial class SchemaInitializationLog {
   [LoggerMessage(
       Level = LogLevel.Debug,
       Message = "Acquiring advisory lock {LockId} for schema '{Schema}'...")]
-  public static partial void AcquiringAdvisoryLock(ILogger logger, int lockId, string schema);
+  public static partial void AcquiringAdvisoryLock(ILogger logger, long lockId, string schema);
 
   [LoggerMessage(
       Level = LogLevel.Debug,
       Message = "Advisory lock {LockId} held by another instance, retrying (attempt {Attempt}, delay {DelayMs}ms)...")]
-  public static partial void AdvisoryLockRetry(ILogger logger, int lockId, int attempt, int delayMs);
+  public static partial void AdvisoryLockRetry(ILogger logger, long lockId, int attempt, int delayMs);
 
   [LoggerMessage(
       Level = LogLevel.Debug,
       Message = "Acquired advisory lock {LockId} for schema '{Schema}'")]
-  public static partial void AcquiredAdvisoryLock(ILogger logger, int lockId, string schema);
+  public static partial void AcquiredAdvisoryLock(ILogger logger, long lockId, string schema);
 
   [LoggerMessage(
       Level = LogLevel.Debug,
       Message = "Transaction committed, advisory lock {LockId} released for schema '{Schema}'")]
-  public static partial void TransactionCommitted(ILogger logger, int lockId, string schema);
+  public static partial void TransactionCommitted(ILogger logger, long lockId, string schema);
 
   [LoggerMessage(
       Level = LogLevel.Warning,
